Fix/layout issues

+++ Dinesafe/DSFScorebarView.m
@@ -44,9 +44,9 @@ const float kScoreBoxGap = 0; // Gap between boxes
     UIFont *yearFont = [UIFont fontWithName:@"PFTempestaFiveCompressed" size:8.0];
 
     // take subset/slice of inspections. only the last 17, so it will fit on screen.
-    int inspections_count = [self.inspections count];
-    int startIndex = inspections_count > 17 ? inspections_count - 17 - 1 : 0;
-    int subarrayLength = inspections_count > 17 ? 17 : inspections_count;
+    NSUInteger inspections_count = [self.inspections count];
+    NSUInteger startIndex = inspections_count > 17 ? inspections_count - 17 - 1 : 0;
+    NSUInteger subarrayLength = inspections_count > 17 ? 17 : inspections_count;
     NSArray *inspectionsSlice = [self.inspections subarrayWithRange:NSMakeRange(startIndex, subarrayLength)];
 
     for (id inspection in inspectionsSlice) {
